+++ src/main/java/com/allclear/tastytrack/domain/restaurant/service/RawDataService.java
@@ -1,20 +1,15 @@
 package com.allclear.tastytrack.domain.restaurant.service;
 
-import com.allclear.tastytrack.domain.restaurant.coordinate.dto.Coordinate;
-import com.allclear.tastytrack.domain.restaurant.coordinate.service.CoordinateService;
 import com.allclear.tastytrack.domain.restaurant.dto.LocalDataResponse;
 import com.allclear.tastytrack.domain.restaurant.dto.RawRestaurantResponse;
 import com.allclear.tastytrack.domain.restaurant.entity.RawRestaurant;
-import com.allclear.tastytrack.domain.restaurant.entity.Restaurant;
 import com.allclear.tastytrack.domain.restaurant.repository.RawRestaurantRepository;
-import com.allclear.tastytrack.domain.restaurant.repository.RestaurantRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -30,13 +25,12 @@ import java.util.List;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class ApiService {
+public class RawDataService {
 
     private final RawRestaurantRepository rawRestaurantRepository;
-    private final RestaurantRepository restaurantRepository;
     private final ObjectMapper objectMapper;
 
-    private final CoordinateService coordinateService;
+    private final DataProcessingService dataProcessingService;
 
     @Value("${API_URL}")
     private String apiUrl; // API URL
@@ -73,7 +67,7 @@ public class ApiService {
         fetchAndSaveCommon();
 
         // 최초 맛집 가공 DB 저장
-        preprocessingAndSaveInitRestaurant();
+        dataProcessingService.preprocessingAndSaveInitRestaurant();
     }
 
     /**
@@ -89,7 +83,7 @@ public class ApiService {
         fetchAndSaveCommon();
 
         // 업데이트된 데이터만 맛집 가공 DB 저장
-        preprocessingAndSaveUpdatedAtNotMatchingRestaurant();
+        dataProcessingService.preprocessingAndSaveUpdatedAtNotMatchingRestaurant();
     }
 
     /**
@@ -184,35 +178,6 @@ public class ApiService {
     }
 
     /**
-     * 맛집 원본 데이터 전처리 후 가공된 데이터를 맛집 가공 테이블에 저장하여 초기 데이터를 구축합니다.
-     * 작성자 : 유리빛나
-     */
-    public void preprocessingAndSaveInitRestaurant() throws Exception {
-
-        // 맛집 원본 리스트의 모든 데이터 조회
-        List<RawRestaurant> rawRestaurantList = rawRestaurantRepository.findAll();
-
-        // 맛집 원본 리스트를 맛집 가공 DB에 저장
-        saveRestaurantsFromRawRestaurants(rawRestaurantList);
-    }
-
-    /**
-     * 맛집 원본과 맛집 가공의 최종수정일자가 다른 맛집 원본 데이터 전처리 후 가공된 데이터를 맛집 가공 테이블에 저장합니다.
-     * 작성자 : 유리빛나
-     */
-    public void preprocessingAndSaveUpdatedAtNotMatchingRestaurant() throws Exception {
-
-        // 맛집 원본과 맛집 가공의 최종수정일자가 다른 맛집 원본 테이블의 데이터 조회
-        List<RawRestaurant> rawRestaurantList = rawRestaurantRepository.findByLastUpdatedAtNotMatchingRawRestaurants();
-
-        // 업데이트된 맛집 원본 데이터가 없을 경우 리턴
-        if (rawRestaurantList.isEmpty()) return;
-
-        // 맛집 원본 리스트를 맛집 가공 DB에 저장
-        saveRestaurantsFromRawRestaurants(rawRestaurantList);
-    }
-
-    /**
      * 공공데이터의 맛집 응답 데이터를 맛집 원본 Entity에 저장하기 위해 Builder 패턴으로 RawRestaurant 객체를 생성하는 메서드
      * 작성자 : 유리빛나
      *
@@ -247,80 +212,6 @@ public class ApiService {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 
         return LocalDateTime.parse(lastmodts, formatter);
-    }
-
-    /**
-     * 맛집 원본 데이터를 맛집 가공 Entity에 저장하기 위해 Builder 패턴으로 Restaurant 객체를 생성하는 메서드
-     * 작성자 : 유리빛나
-     *
-     * @param rawRestaurant 맛집 원본 데이터
-     * @return 맛집 원본 데이터가 저장된 맛집 가공 Entity 객체
-     */
-    private Restaurant getRestaurantBuilder(RawRestaurant rawRestaurant) {
-
-        return Restaurant.builder()
-                .code(rawRestaurant.getMgtno())
-                .name(rawRestaurant.getBplcnm())
-                .type(rawRestaurant.getUptaenm())
-                .status(rawRestaurant.getDtlstategbn())
-                .oldAddress(rawRestaurant.getSitewhladdr())
-                .newAddress(rawRestaurant.getRdnwhladdr())
-                .lon(Double.valueOf(rawRestaurant.getLon()))
-                .lat(Double.valueOf(rawRestaurant.getLat()))
-                .lastUpdatedAt(parseLastmodts(rawRestaurant.getLastmodts()))
-                .build();
-    }
-
-    /**
-     * 맛집 원본 리스트를 맛집 가공 DB에 저장합니다.
-     * 작성자 : 유리빛나
-     *
-     * @param rawRestaurantList 맛집 가공 DB에 저장할 맛집 원본 리스트
-     */
-    private void saveRestaurantsFromRawRestaurants(List<RawRestaurant> rawRestaurantList) throws Exception {
-
-        for (RawRestaurant rawRestaurant : rawRestaurantList) {
-            // 도로명주소가 없는 데이터는 일단 제외
-            if (rawRestaurant.getRdnwhladdr().isEmpty()) continue;
-
-            // 위도, 경도 값이 누락된 데이터 주입
-            if (rawRestaurant.getLon().isEmpty()) {
-                Coordinate coordinate = coordinateService.getCoordinate(rawRestaurant.getRdnwhladdr());
-                rawRestaurant.setLon(coordinate.getLon());
-                rawRestaurant.setLat(coordinate.getLat());
-            }
-
-            // 가공 테이블에서 동일한 mgtno(code) 값을 가진 데이터를 조회
-            Restaurant existingRestaurant = restaurantRepository.findByCode(rawRestaurant.getMgtno());
-
-            if (existingRestaurant != null) {
-                // 기존 데이터가 있는 경우, 최종 수정일자를 비교하여 업데이트
-                LocalDateTime existingLastUpdatedAt = existingRestaurant.getLastUpdatedAt();
-                LocalDateTime newLastUpdatedAt = parseLastmodts(rawRestaurant.getLastmodts());
-
-                if (!existingLastUpdatedAt.equals(newLastUpdatedAt)) {
-                    // 최종수정일자가 다른 경우에만 업데이트
-                    existingRestaurant.updateWithNewData(getRestaurantBuilder(rawRestaurant));
-                    restaurantRepository.save(existingRestaurant);
-                }
-            } else {
-                // 신규 데이터인 경우에만 저장
-                Restaurant restaurant = getRestaurantBuilder(rawRestaurant);
-
-                // 맛집 원본에 폐업일자가 있는 데이터는 맛집 가공의 삭제여부를 true로 저장
-                if (!rawRestaurant.getDcbymd().isEmpty()) {
-                    restaurant.setDeletedYn(true);
-                }
-
-                try {
-                    // 가공된 데이터 저장
-                    restaurantRepository.save(restaurant);
-                } catch (DataIntegrityViolationException e) {
-                    log.error("중복된 데이터로 인해 저장 실패: {}", rawRestaurant.getMgtno(), e);
-                    // 중복된 데이터로 인한 예외가 발생한 경우 로깅 및 처리
-                }
-            }
-        }
     }
 
 }
